Move WriteFile logic to util package
